Volume property type changed for crypto;

+++ 29_Financial/Trial_07_OHLCV_Spike/List_Comparer/OhlcvObject.cs
@@ -22,7 +22,7 @@ namespace List_Comparer
         public DateTime TimeStamp { get; set; }
 
         [JsonProperty(PropertyName = "volume")]
-        public int Volume { get; set; }
+        public decimal Volume { get; set; }
         public int Index { get; set; }
         public decimal Capital { get; set; }
         public decimal Multiplier { get; set; }
